Course_Importer/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
google-api-python-client==2.125.0
google-auth==2.29.0
gspread==6.1.2
python-dotenv==1.0.1
pydantic==2.7.1